BaseWindow: We are now grabbing the mouse. Doing this because on linux, you need to grab and move the pointer or else you can scroll away from your full screen program.
org_lwjgl_input_Mouse.cpp: When in full screen, the pointer should warp
to the corner of the display. This fixes a few issues.

Also added a stub for nEnableBuffer, with a fixme message. This function
does not seem to be implemented in either the linux or win32 source...
but exists in the java Mouse class. The stub keeps programs from crashing.

ltmain.sh: Fixed a bug in libtool, where it uses gcc vs c++ for linking
librarys built in c++. This fixes a missing symbol issue.

configure.in: Made a few small changes.